+++ include/engine/Texture.hpp
@@ -45,14 +45,14 @@ namespace Guarneri
 		TextureFormat format;
 		std::string raw_path;
 		std::string meta_path;
-		uint32_t width;
-		uint32_t height;
-		uint32_t mip_count;
+		size_t width;
+		size_t height;
+		size_t mip_count;
 		Filtering mip_filtering;
 		bool enable_mip;
 
 	private:
-		static std::unordered_map<uint32_t, std::shared_ptr<Texture>> texture_cache;
+		static std::unordered_map<size_t, std::shared_ptr<Texture>> texture_cache;
 		std::shared_ptr<RawBuffer<tinymath::color_rgb16f>> rgb16f_buffer;
 		std::shared_ptr<RawBuffer<tinymath::color_rgba16f>> rgba16f_buffer;
 		std::shared_ptr<RawBuffer<tinymath::color_rgb>> rgb_buffer;
@@ -67,8 +67,8 @@ namespace Guarneri
 		std::vector< std::shared_ptr<RawBuffer<tinymath::color_rgba16f>>> rgba16f_mipmaps;
 
 	public:
-		Texture(const uint32_t& _width, const uint32_t& _height, const TextureFormat& _fmt);
-		Texture(void* tex_buffer, const uint32_t& _width, const uint32_t& _height, const TextureFormat& _fmt);
+		Texture(const size_t& _width, const size_t& _height, const TextureFormat& _fmt);
+		Texture(void* tex_buffer, const size_t& _width, const size_t& _height, const TextureFormat& _fmt);
 		Texture(const Texture& other);
 		~Texture();
 
@@ -78,17 +78,17 @@ namespace Guarneri
 		static std::shared_ptr<Texture> load_raw(const char* path);
 
 		void reload(const char* texture_path);
-		bool bilinear(const float& u, const float& v, const uint32_t& mip, tinymath::Color& ret) const;
-		bool point(const float& u, const float& v, const uint32_t& mip, tinymath::Color& ret) const;
+		bool bilinear(const float& u, const float& v, const size_t& mip, tinymath::Color& ret) const;
+		bool point(const float& u, const float& v, const size_t& mip, tinymath::Color& ret) const;
 		void generate_mipmap(const int& mip_count, const Filtering& filtering);
 		bool sample(const float& u, const float& v, tinymath::Color& ret) const;
-		bool sample(const float& u, const float& v, const uint32_t& mip, tinymath::Color& ret) const;
+		bool sample(const float& u, const float& v, const size_t& mip, tinymath::Color& ret) const;
 		bool sample(const float& u, const float& v, const float& lod, tinymath::Color& ret) const;
-		bool read(const float& u, const float& v, const uint32_t& mip, tinymath::Color& ret) const;
-		bool read(const uint32_t& row, const uint32_t& col, const uint32_t& mip, tinymath::Color& ret) const;
+		bool read(const float& u, const float& v, const size_t& mip, tinymath::Color& ret) const;
+		bool read(const size_t& row, const size_t& col, const size_t& mip, tinymath::Color& ret) const;
 		bool read(const float& u, const float& v, tinymath::Color& ret) const;
-		bool read(const uint32_t& row, const uint32_t& col, tinymath::Color& ret) const;
-		bool write(const uint32_t& x, const uint32_t& y, const tinymath::Color& data);
+		bool read(const size_t& row, const size_t& col, tinymath::Color& ret) const;
+		bool write(const size_t& x, const size_t& y, const tinymath::Color& data);
 		bool write(const float& u, const float& v, const tinymath::Color& data);
 		void save2file();
 		void resize();
