+++ include/rasterizor/FrameTile.hpp
@@ -8,10 +8,10 @@ namespace Guarneri
 {
 	typedef struct
 	{
-		uint32_t tile_size;
-		uint32_t row_tile_count;
-		uint32_t col_tile_count;
-		uint32_t tile_length;
+		size_t tile_size;
+		size_t row_tile_count;
+		size_t col_tile_count;
+		size_t tile_length;
 	} TileInfo;
 
 	struct TileTask
@@ -28,11 +28,11 @@ namespace Guarneri
 	struct FrameTile
 	{
 	public:
-		uint32_t tile_idx;
-		int row_start;
-		int row_end;
-		int col_start;
-		int col_end;
+		size_t tile_idx;
+		size_t row_start;
+		size_t row_end;
+		size_t col_start;
+		size_t col_end;
 		SafeQueue<TileTask> tasks;
 
 	public:
@@ -42,20 +42,20 @@ namespace Guarneri
 		bool pop_task(TileTask& task);
 		bool is_task_empty();
 		size_t task_size();
-		static int coord2index(const int& row, const int& col, const int& col_tile_count);
-		static void pixel2tile(const int& prow, const int& pcol, int& trow, int& tcol, const int& tile_size);
+		static size_t coord2index(const size_t& row, const size_t& col, const size_t& col_tile_count);
+		static void pixel2tile(const size_t& prow, const size_t& pcol, size_t& trow, size_t& tcol, const size_t& tile_size);
 		static void build_tiles(
 			FrameTile* tiles,
-			const int& tile_size,
-			const int& row_tile_count, const int& col_tile_count,
-			const int& row_rest, const int& col_rest);
+			const size_t& tile_size,
+			const size_t& row_tile_count, const size_t& col_tile_count,
+			const size_t& row_rest, const size_t& col_rest);
 		static void push_draw_task(
 			FrameTile* tiles,
-			const uint32_t& row_tile_count,
-			const uint32_t& col_tile_count,
+			const size_t& row_tile_count,
+			const size_t& col_tile_count,
 			const Triangle& tri,
 			const Shader& shader,
-			const int& w, const int& h,
-			const int& tile_size);
+			const size_t& w, const size_t& h,
+			const size_t& tile_size);
 	};
 }
