+++ include/rasterizor/RawBuffer.hpp
@@ -12,29 +12,29 @@ namespace Guarneri
 		void (*deletor)(T* ptr);
 
 	public:
-		uint32_t width;
-		uint32_t height;
+		size_t width;
+		size_t height;
 
 	public:
-		RawBuffer(uint32_t _width, uint32_t _height);
-		RawBuffer(void* _buffer, uint32_t _width, uint32_t _height, void (*deletor)(T* ptr));
+		RawBuffer(size_t _width, size_t _height);
+		RawBuffer(void* _buffer, size_t _width, size_t _height, void (*deletor)(T* ptr));
 		RawBuffer(const RawBuffer<T>& other);
 		~RawBuffer();
-		static std::shared_ptr<RawBuffer> create(uint32_t _width, uint32_t _height);
-		static std::shared_ptr<RawBuffer> create(void* _buffer, uint32_t _width, uint32_t _height, void (*deletor)(T* ptr));
+		static std::shared_ptr<RawBuffer> create(size_t _width, size_t _height);
+		static std::shared_ptr<RawBuffer> create(void* _buffer, size_t _width, size_t _height, void (*deletor)(T* ptr));
 		static std::shared_ptr<RawBuffer> create(const RawBuffer<T>& other);
 		bool read(const float& u, const float& v, T& out) const;
-		bool read(const uint32_t& row, const uint32_t& col, T& out) const;
+		bool read(const size_t& row, const size_t& col, T& out) const;
 		bool write(const float& u, const float& v, const T& data);
-		bool write(const uint32_t& row, const uint32_t& col, const T& data);
+		bool write(const size_t& row, const size_t& col, const T& data);
 		void clear(const T& val);
 		T* get_ptr(size_t& size);
 		RawBuffer<T>& operator = (const RawBuffer<T>& other);
 		void copy(const RawBuffer<T>& other);
 	};
 
-	void uv2pixel(const uint32_t& width, const uint32_t& height, const float& u, const float& v, uint32_t& row, uint32_t& col);
-	void pixel2uv(const uint32_t& width, const uint32_t& height, const uint32_t& row, const uint32_t& col, float& u, float& v);
+	void uv2pixel(const size_t& width, const size_t& height, const float& u, const float& v, size_t& row, size_t& col);
+	void pixel2uv(const size_t& width, const size_t& height, const size_t& row, const size_t& col, float& u, float& v);
 }
 
 #include "detail/RawBuffer.inl"
